# docs/swagger.yml
swagger: '2.0'
info:
  title: Blackfury Chain - REST API
  description: REST interface for query and transaction services
  version: 1.0.0
paths:
...
  /blackfury/claimsmanager/v1/previous_epoch_claims/{chain_id}:
    get:
      summary: LastEpochClaims returns all zone claims from the last epoch.
      operationId: LastEpochClaims
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              claims:
                type: array
                items:
                  type: object
                  properties:
                    user_address:
                      type: string
                    chain_id:
                      type: string
                    module:
                      type: string
                      enum:
                        - ClaimTypeUndefined
                        - ClaimTypeLiquidToken
                        - ClaimTypeOsmosisPool
                        - ClaimTypeCrescentPool
                        - ClaimTypeSifchainPool
                      default: ClaimTypeUndefined
                      title: >-
                        - ClaimTypeUndefined: Undefined action (per protobuf
                        spec)
                    source_chain_id:
                      type: string
                    amount:
                      type: string
                      format: uint64
                  description: >-
                    Claim define the users claim for holding assets within a
                    given zone.
              pagination:
                type: object
                properties:
                  next_key:
                    type: string
                    format: byte
                    description: |-
                      next_key is the key to be passed to PageRequest.key to
                      query the next page most efficiently. It will be empty if
                      there are no more results.
                  total:
                    type: string
                    format: uint64
                    title: >-
                      total is total number of results available if
                      PageRequest.count_total

                      was set, its value is undefined otherwise
                description: >-
                  PageResponse is to be embedded in gRPC response messages where
                  the

                  corresponding request message has used PageRequest.

                   message SomeResponse {
                           repeated Bar results = 1;
                           PageResponse page = 2;
                   }
            description: >-
              QueryClaimsResponse is the response type for the Query/Claims RPC
              method.
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: chain_id
          in: path
          required: true
          type: string
        - name: address
          in: query
          required: false
          type: string
        - name: pagination.key
          description: |-
            key is a value returned in PageResponse.next_key to begin
            querying the next page most efficiently. Only one of offset or key
            should be set.
          in: query
          required: false
          type: string
          format: byte
        - name: pagination.offset
          description: >-
            offset is a numeric offset that can be used when key is unavailable.

            It is less efficient than using key. Only one of offset or key
            should

            be set.
          in: query
          required: false
          type: string
          format: uint64
        - name: pagination.limit
          description: >-
            limit is the total number of results to be returned in the result
            page.

            If left empty it will default to a value to be set by each app.
          in: query
          required: false
          type: string
          format: uint64
        - name: pagination.count_total
          description: >-
            count_total is set to true  to indicate that the result set should
            include

            a count of the total number of items available for pagination in
            UIs.

            count_total is only respected when offset is used. It is ignored
            when key

            is set.
          in: query
          required: false
          type: boolean
        - name: pagination.reverse
          description: >-
            reverse is set to true if results are to be returned in the
            descending order.


            Since: cosmos-sdk 0.43
          in: query
          required: false
          type: boolean
      tags:
        - QueryClaimsManager
  /blackfury/claimsmanager/v1/user/{address}/claims:
    get:
      summary: >-
        UserClaims returns all zone claims for a given address from the current
        epoch.
      operationId: UserClaims
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              claims:
                type: array
                items:
                  type: object
                  properties:
                    user_address:
                      type: string
                    chain_id:
                      type: string
                    module:
                      type: string
                      enum:
                        - ClaimTypeUndefined
                        - ClaimTypeLiquidToken
                        - ClaimTypeOsmosisPool
                        - ClaimTypeCrescentPool
                        - ClaimTypeSifchainPool
                      default: ClaimTypeUndefined
                      title: >-
                        - ClaimTypeUndefined: Undefined action (per protobuf
                        spec)
                    source_chain_id:
                      type: string
                    amount:
                      type: string
                      format: uint64
                  description: >-
                    Claim define the users claim for holding assets within a
                    given zone.
              pagination:
                type: object
                properties:
                  next_key:
                    type: string
                    format: byte
                    description: |-
                      next_key is the key to be passed to PageRequest.key to
                      query the next page most efficiently. It will be empty if
                      there are no more results.
                  total:
                    type: string
                    format: uint64
                    title: >-
                      total is total number of results available if
                      PageRequest.count_total

                      was set, its value is undefined otherwise
                description: >-
                  PageResponse is to be embedded in gRPC response messages where
                  the

                  corresponding request message has used PageRequest.

                   message SomeResponse {
                           repeated Bar results = 1;
                           PageResponse page = 2;
                   }
            description: >-
              QueryClaimsResponse is the response type for the Query/Claims RPC
              method.
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: address
          in: path
          required: true
          type: string
        - name: chain_id
          in: query
          required: false
          type: string
        - name: pagination.key
          description: |-
            key is a value returned in PageResponse.next_key to begin
            querying the next page most efficiently. Only one of offset or key
            should be set.
          in: query
          required: false
          type: string
          format: byte
        - name: pagination.offset
          description: >-
            offset is a numeric offset that can be used when key is unavailable.

            It is less efficient than using key. Only one of offset or key
            should

            be set.
          in: query
          required: false
          type: string
          format: uint64
        - name: pagination.limit
          description: >-
            limit is the total number of results to be returned in the result
            page.

            If left empty it will default to a value to be set by each app.
          in: query
          required: false
          type: string
          format: uint64
        - name: pagination.count_total
          description: >-
            count_total is set to true  to indicate that the result set should
            include

            a count of the total number of items available for pagination in
            UIs.

            count_total is only respected when offset is used. It is ignored
            when key

            is set.
          in: query
          required: false
          type: boolean
        - name: pagination.reverse
          description: >-
            reverse is set to true if results are to be returned in the
            descending order.


            Since: cosmos-sdk 0.43
          in: query
          required: false
          type: boolean
      tags:
        - QueryClaimsManager
...